Example of test service for JWT Verification
This is an example of how a test service verifies a JWT Token with Clerk.

Component `jwttestui.tsx` uses end point by test-service to verify the JWT token.

You can try it out by retrieving the token in the frontend and storing it in localStorage where the frontend calls the test-service endpoint to verify the token. Valid verification will be reflected on the UI and otherwise.